+++ src/app/@modal/(.)temp/page.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import { EmailLoginForm } from '@/app/signIn/EmailLoginForm';
-import { LoginHeader } from '@/app/signIn/LoginHeader';
-import { SocialLogin } from '@/app/signIn/SocialLogin';
+import { EmailLoginForm } from '@/app/temp/EmailLoginForm';
+import { LoginHeader } from '@/app/temp/LoginHeader';
+import { SocialLogin } from '@/app/temp/SocialLogin';
 
 export default function LoginModal() {
   const router = useRouter();
